requirements.txt
streamlit
google-generativeai
pypdf
duckduckgo-search
edge-tts
requests